rf3: don't upcast the MLFF constant -- it breaks the bit-exactness it was added for
The previous commit claimed C_L bit-exact and, as committed, it was not: adding
.float() to mlff_constant's return promoted the addition. Under autocast the
reference adds a bf16 constant to a bf16 C_L; handing back an fp32 copy makes the
sum fp32 and lands somewhere else. C_L went back to pcc 0.99999-ish -- close enough
to read as fine, which is exactly why the probe checks bit-exactness and not just
pcc.

I introduced it in the same commit that verified the result, by editing the helper
after running the probe and re-running only to see PASS, which the verdict still
said because the pcc thresholds were met. The check that caught it was the
per-tensor bit_exact flag printed beside the verdict.

Constant is now returned in the dtype the reference produces it in. The cast to
fp32 belongs at the device boundary, where ttnn does it anyway.

    P_LL   bit-exact   pcc 1.0, maxabs 0.0
    C_L    bit-exact   pcc 1.0, maxabs 0.0     (rna, L=115)

+++ tt_bio/rf3/feature_init.py
@@ -138,7 +138,11 @@ def mlff_constant(process_atom_level_embedding, n_conformers: int = 8,
         out = process_atom_level_embedding(
             torch.zeros(n_conformers, 1, embedding_dim)
         )
-    return out[0].float().clone()
+    # Returned in the dtype the reference produces it in, NOT upcast: under autocast
+    # the reference adds a bf16 constant to a bf16 C_L, and handing back an fp32 copy
+    # promotes that addition and changes the result. Cast at the device boundary
+    # instead, where ttnn does it anyway.
+    return out[0].clone()
 
 
 def assert_mlff_inputs_zero(f: dict) -> None:
